% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{conv_fpath_to_uri_fpath}
\alias{conv_fpath_to_uri_fpath}
\title{convert file path to URI-compatible file path
on OS'es with file separator different that "/" convert file separator to "/"}
\usage{
conv_fpath_to_uri_fpath(fpath)
}
\arguments{
\item{fpath}{string with file path}
}
\value{
string with URI-compatible file path
}
\description{
convert file path to URI-compatible file path
on OS'es with file separator different that "/" convert file separator to "/"
}
