% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CramTrack.R
\docType{class}
\name{CramTrack-class}
\alias{CramTrack-class}
\alias{.CramTrack}
\alias{CramTrack}
\title{Constructor for CramTrack}
\usage{
CramTrack(
  trackName,
  cramUrl,
  indexUrl,
  trackHeight = 50,
  visibilityWindow = 30000,
  color = "gray"
)
}
\arguments{
\item{trackName}{A character string, used as track label by igv.}

\item{cramUrl}{The URL of a cram file.}

\item{indexUrl}{The URL of a cram index file (.crai).}

\item{trackHeight}{track height, typically in range 20 (for annotations) and up to 1000.}

\item{visibilityWindow}{Maximum window size in base pairs for which alignments are displayed. Defaults to 30 kb.}

\item{color}{A character string, either a recognized color ("red") or a hex string.}
}
\value{
A CramTrack object
}
\description{
\code{CramTrack} creates an \code{IGV} track for remote CRAM files
}
