% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeSpec.R
\name{currently.supported.stock.genomes}
\alias{currently.supported.stock.genomes}
\title{currently.supported.stock.genomes}
\usage{
currently.supported.stock.genomes(test = FALSE)
}
\value{
an list of short genome codes, e.g., "hg38", "dm6", "tair10"
}
\description{
a helper function for mostly internal use, obtains the genome codes (e.g. 'hg38')
      supported by igv.js
}
