% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicAlignmentsTrack.R
\name{GenomicAlignmentsTrack}
\alias{GenomicAlignmentsTrack}
\alias{GenomicAlignmentTrack}
\title{Constructor for GenomicAlignmentsTrack}
\usage{
GenomicAlignmentTrack(trackName, alignment, trackHeight = 50,
  visibilityWindow = 30000, color = "gray")
}
\arguments{
\item{trackName}{A character string, used as track label by igv, we recommend unique names per track.}

\item{alignment}{A GAlignments object}

\item{trackHeight}{track height, typically in range 20 (for annotations) and up to 1000 (for large sample vcf files)}

\item{visibilityWindow}{Maximum window size in base pairs for which indexed annotations or variants are displayed. Defaults: 1 MB for variants, whole chromosome for other track types.}
}
\value{
A GenomicAlignmentsTrack object
}
\description{
\code{GenomicAlignmentsTrack} creates and \code{IGV} track for bed-like objects expressed as GRanges
}
\details{
Detailed description goes here
}
\examples{

  bamFile <- system.file(package="igvR", "extdata", "tumor.bam")
  which <- GRanges(seqnames = "21", ranges = IRanges(10400126, 10400326))
  param <- ScanBamParam(which=which, what = scanBamWhat())
  x <- readGAlignments(bamFile, use.names=TRUE, param=param)
  track <- GenomicAlignmentTrack("tumor", x)

}
