% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesAnnotationTrack.R
\docType{methods}
\name{getSize,GRangesAnnotationTrack-method}
\alias{getSize,GRangesAnnotationTrack-method}
\title{Retrieve the size of the GRangesAnnotationTrack}
\usage{
\S4method{getSize}{GRangesAnnotationTrack}(obj)
}
\arguments{
\item{obj}{An object of class GRangesAnnotationTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the GRangesAnnotationTrack
}
