% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSCBedAnnotationTrack.R
\docType{methods}
\name{getSize,UCSCBedAnnotationTrack-method}
\alias{getSize,UCSCBedAnnotationTrack-method}
\title{Retrieve the size of theUCSCBedAnnotationTrack}
\usage{
\S4method{getSize}{UCSCBedAnnotationTrack}(obj)
}
\arguments{
\item{obj}{An object of class UCSCBedAnnotationTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of theUCSCBedAnnotationTrack
}
\examples{
bed.filepath <- system.file(package = "rtracklayer", "tests", "test.bed")
gr.bed <- import(bed.filepath)
track.1 <- UCSCBedAnnotationTrack("UCSC bed", gr.bed,  color="blue", displayMode="SQUISHED")
getSize(track.1)

}
