% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantTrack.R
\docType{methods}
\name{getSize,VariantTrack-method}
\alias{getSize,VariantTrack-method}
\title{Retrieve the size of the VariantTrack}
\usage{
\S4method{getSize}{VariantTrack}(obj)
}
\arguments{
\item{obj}{An object of class VariantTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the VariantTrack
}
