% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{getSupportedGenomes,igvR-method}
\alias{getSupportedGenomes,igvR-method}
\alias{getSupportedGenomes}
\title{Get the shorthand codes (eg, "hg38") for the genomes currently supported by our use of igv.js}
\usage{
\S4method{getSupportedGenomes}{igvR}(obj)
}
\arguments{
\item{obj}{An object of class igvR}
}
\value{
A character vector, the short form names of the currently supported genomes
}
\description{
Get the shorthand codes (eg, "hg38") for the genomes currently supported by our use of igv.js
}
\examples{
if (interactive()) {
   igv <- igvR()
   getSupportedGenomes(igv)
   }

}
