% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGV.R
\docType{class}
\name{IGV-class}
\alias{IGV-class}
\alias{.IGV}
\alias{IGV}
\title{Create an IGV object}
\usage{
IGV(portRange = 15000:15100, host = "localhost", title = "IGV",
  quiet = TRUE)
}
\arguments{
\item{portRange}{The constructor looks for a free websocket port in this range.  15000:15100 by default}

\item{host}{In practice, this is always "localhost"}

\item{title}{Used for the web browser window, "IGV" by default}

\item{quiet}{A logical variable controlling verbosity during execution}
}
\value{
An object of the IGV class
}
\description{
The IGV class provides an R interface to igv.js, a rich, interactive, full-featured, javascript
browser-based genome browser.  One constructs an IGV instance on a specified port (default 9000),
the browser code is loaded, and a websocket connection openend.  After specifying the reference
genome, any number of genome tracks may be created, displayed, and navigated.
}
\examples{
if(interactive()){
   igv <- IGV(title="igv demo")
   Sys.sleep(2)
   setGenome(igv, "hg38")
   Sys.sleep(5)
   showGenomicRegion(igv, "MEF2C")
     #---------------------------------------------------------------
     # an easy transparent way to create a bed track
     #---------------------------------------------------------------
   base.loc <- 88883100
   tbl <- data.frame(chrom=rep("chr5", 3),
                     start=c(base.loc, base.loc+100, base.loc + 250),
                     end=c(base.loc + 50, base.loc+120, base.loc+290),
                     name=c("a", "b", "c"),
                     score=runif(3),
                     strand=rep("*", 3),
                     stringsAsFactors=FALSE)

   track <- DataFrameAnnotationTrack("dataframeTest", tbl, color="red", displayMode="EXPANDED")
   displayTrack(igv, track)
   showGenomicRegion(igv, sprintf("chr5:\%d-\%d", base.loc-100, base.loc+350))
   } # if !interactive

}
