% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printf}
\alias{printf}
\title{printf}
\usage{
printf(...)
}
\arguments{
\item{...}{a format string followed by a matching number of scalar values}
}
\value{
a string
}
\description{
\code{printf} emulates the standard C library function of the same name
}
\examples{
  printf("\%s: \%04d, \%10.8f", "emulating standard C library", 10, pi)

}
