% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{saveToSVG,igvR-method}
\alias{saveToSVG,igvR-method}
\alias{saveToSVG}
\title{Get entire igv browser image in svg}
\usage{
\S4method{saveToSVG}{igvR}(obj, filename)
}
\arguments{
\item{obj}{An object of class igvR}

\item{filename}{character string, the name of the file to which the svg text will be written}
}
\value{
A character vector
}
\description{
Get entire igv browser image in svg
}
