% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\docType{methods}
\name{setTrackClickFunction,igvR-method}
\alias{setTrackClickFunction,igvR-method}
\alias{setTrackClickFunction}
\title{Set the visible region, by explicit chromLoc string, or by named features in any curently loaded
annotation tracks}
\usage{
\S4method{setTrackClickFunction}{igvR}(obj, javascriptFunction)
}
\arguments{
\item{obj}{An object of class igvR}

\item{region}{A genomic location (rendered "chr5:9,234,343-9,236,000" or as a list:
list(chrom="chr9", start=9234343, end=9236000)) or a labeled annotation in a searchable track,
often a gene symbol, eg "MEF2C"}
}
\value{
""
}
\description{
Set the visible region, by explicit chromLoc string, or by named features in any curently loaded
annotation tracks
}
