% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{setTrackHeight,igvR-method}
\alias{setTrackHeight,igvR-method}
\alias{setTrackHeight}
\title{Remove named tracks}
\usage{
\S4method{setTrackHeight}{igvR}(obj, trackName, newHeight)
}
\arguments{
\item{obj}{An object of class igvR}

\item{trackName}{a character string}

\item{newHeight}{integer, in ixels}
}
\value{
nothing
}
\description{
Remove named tracks
}
\seealso{
getTrackNames
}
