% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{showTrackLabels,igvR-method}
\alias{showTrackLabels,igvR-method}
\alias{showTrackLabels}
\title{Hide or show igv track labels}
\usage{
\S4method{showTrackLabels}{igvR}(obj, newState)
}
\arguments{
\item{obj}{An object of class igvR}

\item{newState}{logigal, either TRUE or FALSE}
}
\value{
""
}
\description{
Hide or show igv track labels
}
