% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Track.R
\name{trackInfo,Track-method}
\alias{trackInfo,Track-method}
\alias{trackInfo}
\title{Get basic info about a track: its type, file format, source and S4 class name}
\usage{
\S4method{trackInfo}{Track}(obj)
}
\arguments{
\item{obj}{An object of base class Track}
}
\value{
A list with four fields: trackType, fileFormat, source, class name
}
\description{
Get basic info about a track: its type, file format, source and S4 class name
}
