% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BedpeInteractionsTrack.R
\name{trackSize,BedpeInteractionsTrack-method}
\alias{trackSize,BedpeInteractionsTrack-method}
\title{Retrieve the size of the BedpeInteractionsTrack}
\usage{
\S4method{trackSize}{BedpeInteractionsTrack}(obj)
}
\arguments{
\item{obj}{An object of class BedpeInteractionsTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the BedpeInteractionsTrack
}
