% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameQuantitativeTrack.R
\docType{methods}
\name{trackSize,DataFrameQuantitativeTrack-method}
\alias{trackSize,DataFrameQuantitativeTrack-method}
\title{Retrieve the size of the DataFrameQuantitativeTrack}
\usage{
\S4method{trackSize}{DataFrameQuantitativeTrack}(obj)
}
\arguments{
\item{obj}{An object of class DataFrameQuantitativeTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the DataFrameQuantitativeTrack
}
