% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicAlignmentTrack.R
\docType{methods}
\name{trackSize,GenomicAlignmentTrack-method}
\alias{trackSize,GenomicAlignmentTrack-method}
\title{Retrieve the size of the GenomicAlignmentTrack}
\usage{
\S4method{trackSize}{GenomicAlignmentTrack}(obj)
}
\arguments{
\item{obj}{An object of class GenomicAlignmentTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the GenomicAlignmentTrack
}
