% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesAnnotationTrack.R
\name{trackSize,GRangesAnnotationTrack-method}
\alias{trackSize,GRangesAnnotationTrack-method}
\title{Retrieve the size of the GRangesAnnotationTrack}
\usage{
\S4method{trackSize}{GRangesAnnotationTrack}(obj)
}
\arguments{
\item{obj}{An object of class GRangesAnnotationTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the GRangesAnnotationTrack
}
