% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesQuantitativeTrack.R
\docType{methods}
\name{trackSize,GRangesQuantitativeTrack-method}
\alias{trackSize,GRangesQuantitativeTrack-method}
\title{Retrieve the size of the GRangesQuantitativeTrack}
\usage{
\S4method{trackSize}{GRangesQuantitativeTrack}(obj)
}
\arguments{
\item{obj}{An object of class GRangesQuantitativeTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the GRangesQuantitativeTrack
}
