% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASTrack.R
\name{trackSize,GWASTrack-method}
\alias{trackSize,GWASTrack-method}
\title{Retrieve the size of the GWASTrack}
\usage{
\S4method{trackSize}{GWASTrack}(obj)
}
\arguments{
\item{obj}{An object of class GWASTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the GWASTrack
}
