% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASUrlTrack.R
\name{trackSize,GWASUrlTrack-method}
\alias{trackSize,GWASUrlTrack-method}
\title{Retrieve the size of the GWASUrlTrack}
\usage{
\S4method{trackSize}{GWASUrlTrack}(obj)
}
\arguments{
\item{obj}{An object of class GWASUrlTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the GWASUrlTrack
}
