% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSCBedAnnotationTrack.R
\name{trackSize,UCSCBedAnnotationTrack-method}
\alias{trackSize,UCSCBedAnnotationTrack-method}
\title{Retrieve the size of theUCSCBedAnnotationTrack}
\usage{
\S4method{trackSize}{UCSCBedAnnotationTrack}(obj)
}
\arguments{
\item{obj}{An object of class UCSCBedAnnotationTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of theUCSCBedAnnotationTrack
}
\examples{
bed.filepath <- system.file(package = "rtracklayer", "tests", "test.bed")
gr.bed <- rtracklayer::import(bed.filepath)
track.1 <- UCSCBedAnnotationTrack("UCSC bed", gr.bed, color = "blue", displayMode = "SQUISHED")
trackSize(track.1)

}
