% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSCBedGraphQuantitativeTrack.R
\docType{methods}
\name{trackSize,UCSCBedGraphQuantitativeTrack-method}
\alias{trackSize,UCSCBedGraphQuantitativeTrack-method}
\title{Retrieve the size of the UCSCBedGraphQuantitativeTrack}
\usage{
\S4method{trackSize}{UCSCBedGraphQuantitativeTrack}(obj)
}
\arguments{
\item{obj}{An object of class UCSCBedGraphQuantitativeTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the UCSCBedGraphQuantitativeTrack
}
