% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantitativeTrack.R
\docType{methods}
\name{trackSize,QuantitativeTrack-method}
\alias{trackSize,QuantitativeTrack-method}
\alias{trackSize}
\title{Retrieve the size of the QuantitativeTrack}
\usage{
\S4method{trackSize}{QuantitativeTrack}(obj)
}
\arguments{
\item{obj}{An object of class UCSCBedAnnotationTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the QuantitativeTrack
}
