% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{zoomIn,igvR-method}
\alias{zoomIn,igvR-method}
\alias{zoomIn}
\title{zoom the genome view in by a factor of 2}
\usage{
\S4method{zoomIn}{igvR}(obj)
}
\arguments{
\item{obj}{An object of class igvR}
}
\value{
""
}
\description{
zoom the genome view in by a factor of 2
}
