% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{zoomOut,igvR-method}
\alias{zoomOut,igvR-method}
\alias{zoomOut}
\title{zoom the genome view out by a factor of 2}
\usage{
\S4method{zoomOut}{igvR}(obj)
}
\arguments{
\item{obj}{An object of class igvR}
}
\value{
""
}
\description{
zoom the genome view out by a factor of 2
}
