% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGV.R
\docType{methods}
\name{displayTrack,IGV-method}
\alias{displayTrack,IGV-method}
\alias{displayTrack}
\title{display the specified track in igv}
\usage{
\S4method{displayTrack}{IGV}(obj, track)
}
\arguments{
\item{obj}{An object of class IGV}

\item{track}{An object of some terminal (leaf) subclass of Track}
}
\value{
""
}
\description{
display the specified track in igv
}
\examples{
igv <- IGV()
Sys.sleep(2)
setGenome(igv, "hg38")
Sys.sleep(5)
showGenomicRegion(igv, "MEF2C")
base.loc <- 88883100
tbl <- data.frame(chrom=rep("chr5", 3),
                   start=c(base.loc, base.loc+100, base.loc + 250),
                   end=c(base.loc + 50, base.loc+120, base.loc+290),
                   name=c("a", "b", "c"),
                   score=runif(3),
                   strand=rep("*", 3),
                   stringsAsFactors=FALSE)

track <- DataFrameAnnotationTrack("dataframeTest", tbl, color="red", displayMode="EXPANDED")
displayTrack(igv, track)
}
