% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\docType{methods}
\name{getGenomicRegion,igvR-method}
\alias{getGenomicRegion,igvR-method}
\alias{getGenomicRegion}
\title{Obtain the chromosome and coordiates of the currently display genomic region.}
\usage{
\S4method{getGenomicRegion}{igvR}(obj)
}
\arguments{
\item{obj}{An object of class igvR}
}
\value{
A list with four fields: chrom (character), start(numeric), end(numeric), string(character)
}
\description{
Obtain the chromosome and coordiates of the currently display genomic region.
}
\examples{
if(interactive()){
   igv <- igvR()
   Sys.sleep(2)
   setGenome(igv, "hg38")
   Sys.sleep(5)
   showGenomicRegion(igv, "MEF2C")
   getGenomicRegion(igv)
     # list(chrom="chr5", start=88717241, end=88884466, string="chr5:88,717,241-88,884,466")
   }

}
