% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGV.R
\docType{methods}
\name{ping,IGV-method}
\alias{ping,IGV-method}
\alias{ping}
\title{Test the connection between your R session and the webapp}
\usage{
\S4method{ping}{IGV}(obj)
}
\arguments{
\item{obj}{An object of class IGV}
}
\value{
"pong"
}
\description{
Test the connection between your R session and the webapp
}
\examples{
if(interactive()){
   igv <- IGV()
   ping(igv)
   }
}
