% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\docType{methods}
\name{ping,igvR-method}
\alias{ping,igvR-method}
\alias{ping}
\title{Test the connection between your R session and the webapp}
\usage{
\S4method{ping}{igvR}(obj, msecDelay = 0)
}
\arguments{
\item{obj}{An object of class igvR}

\item{msecDelay}{don't return until these many milliseconds have passed, default 0}
}
\value{
"pong"
}
\description{
Test the connection between your R session and the webapp
}
\examples{
if(webBrowserAvailableForTesting()){
   igv <- igvR()
   ping(igv)
   }
}
