% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGV.R
\docType{methods}
\name{setGenome,IGV-method}
\alias{setGenome,IGV-method}
\alias{setGenome}
\title{Specify the reference genome, currently limited to hg38, hg19, mm10, tair10.}
\usage{
\S4method{setGenome}{IGV}(obj, genomeName)
}
\arguments{
\item{obj}{An object of class IGV}

\item{genomeName}{A character string, one of "hg38", "hg19", "mm10", "tair10"}
}
\value{
An empty string, an error message if the requested genome is not yet supported
}
\description{
Specify the reference genome, currently limited to hg38, hg19, mm10, tair10.
}
\examples{
igv <- IGV()
Sys.sleep(2)
setGenome(igv, "mm10")

}
