% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{setTrackClickFunction,igvR-method}
\alias{setTrackClickFunction,igvR-method}
\alias{setTrackClickFunction}
\title{Specify (supply) the javascript function run on track click event}
\usage{
\S4method{setTrackClickFunction}{igvR}(obj, javascriptFunction)
}
\arguments{
\item{obj}{An object of class igvR}

\item{javascriptFunction}{expressed as a 2-element named list: body + args}
}
\value{
""
}
\description{
Specify (supply) the javascript function run on track click event
}
