% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{setTrackHeight,igvR-method}
\alias{setTrackHeight,igvR-method}
\alias{setTrackHeight}
\title{Remove named tracks}
\usage{
\S4method{setTrackHeight}{igvR}(obj, trackName, newHeight)
}
\arguments{
\item{obj}{An object of class igvR}

\item{trackName}{a character string}

\item{newHeight}{integer, in ixels}
}
\value{
nothing
}
\description{
Remove named tracks
}
\examples{
if(interactive()){
   igv <- igvR()
   setGenome(igv, "hg19")
   showGenomicRegion(igv, "MEF2C")
   setTrackHeigth("Refseq Genes", 100)
   }
}
\seealso{
getTrackNames
}
