% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFF3Track.R
\name{trackSize,GFF3Track-method}
\alias{trackSize,GFF3Track-method}
\title{Retrieve the size of the GFF3Track}
\usage{
\S4method{trackSize}{GFF3Track}(obj)
}
\arguments{
\item{obj}{An object of class UCSCBedAnnotationTrack}
}
\value{
The number of elements
}
\description{
Retrieve the size of the GFF3Track
}
