% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igvR.R
\name{url.exists}
\alias{url.exists}
\title{url.exists}
\usage{
url.exists(url)
}
\arguments{
\item{url}{character the http address to test}
}
\value{
logical TRUE or FALSE
}
\description{
a helper function for mostly internal use, tests for availability of a url,
             modeled after file.exists
}
\examples{

if(interactive()){
   igv <- igvR()
   ping(igv)
   }
}
